refactoring + email shown in text cell
